Implement component property ingest for legacy BOM processing task
